* engine/h2shared/r_draw.c (R_RenderFace): Fixed -Wsign-compare. * engine/hexenworld/server/sv_main.c (SV_SendBan): Changed type from   char to byte to fix -Wnarrowing. * hw_utils/hwmaster/sys_main.c (Sys_ConsoleInput): -Wsign-compare. * hw_utils/hwrcon/hwterm.c (main): Fixed another -Wsign-compare. * libs/timidity/common.c (safe_malloc): Fixed -Wformat in debug msg. * utils/texutils/bsp2wal/bsp2wal.c: De-constified WriteWALFile() and   MakeWorkPath() because they actually write to their string args. * updated version date stamp to 2012-05-16.
git-svn-id: file:///svn/uhexen2/code/trunk@4678 eaf649e6-6c13-0410-b194-84221022858e

+++ engine/h2shared/r_draw.c
@@ -426,7 +426,7 @@ void R_RenderFace (msurface_t *fa, int clipflags)
 			{
 				if (r_pedge->cachededgeoffset & FULLY_CLIPPED_CACHED)
 				{
-					if ((r_pedge->cachededgeoffset & FRAMECOUNT_MASK) == r_framecount)
+					if ((int)(r_pedge->cachededgeoffset & FRAMECOUNT_MASK) == r_framecount)
 					{
 						r_lastvertvalid = false;
 						continue;
@@ -467,7 +467,7 @@ void R_RenderFace (msurface_t *fa, int clipflags)
 			{
 				if (r_pedge->cachededgeoffset & FULLY_CLIPPED_CACHED)
 				{
-					if ((r_pedge->cachededgeoffset & FRAMECOUNT_MASK) == r_framecount)
+					if ((int)(r_pedge->cachededgeoffset & FRAMECOUNT_MASK) == r_framecount)
 					{
 						r_lastvertvalid = false;
 						continue;
